# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CurificBackend",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "http://localhost:5001/api/v1",
			"variables": {}
		}
	],
	"paths": {
...
		"/speciality/listAllSpeciality": {
			"get": {
				"tags": ["Speciality"],
				"summary": "list All speciality",
				"operationId": "listAllspeciality",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2002"
								},
								"example": {
									"success": true,
									"data": [
										{
											"image": "appointment_book/1655721158301-img.png",
											"status": "Available",
											"deletedAt": null,
											"deletedBy": null,
											"_id": "62b046d2d5debb823497dc1c",
											"name": " Back Pain ",
											"price": 700,
											"time": "9:00 AM to 5:00 PM",
											"createdAt": "2022-06-20T10:07:15.090Z",
											"updatedAt": "2022-06-20T10:32:39.125Z",
											"__v": 0
										},
										{
											"image": "appointment_book/1657696907620-image (1).png",
											"status": "Available",
											"deletedAt": null,
											"deletedBy": null,
											"_id": "62ce726328b7481c283e1c77",
											"name": " Back Pain ",
											"price": 700,
											"time": "9:00 AM to 5:00 PM",
											"createdAt": "2022-07-13T07:21:07.024Z",
											"updatedAt": "2022-07-13T07:21:48.467Z",
											"__v": 0
										}
									],
									"totalCount": 2
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/speciality/delete/629df3f75f1a1726a04ccf2f": {
			"delete": {
				"tags": ["Speciality"],
				"summary": "delete speciality",
				"operationId": "deletespeciality",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2003"
								},
								"example": {
									"success": true,
									"msg": "Speciality delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/timeSlot/create": {
			"post": {
				"tags": ["Time Slot"],
				"summary": "Create time Slot",
				"operationId": "CreatetimeSlot",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatetimeSlotRequest"
							},
							"example": {
								"startTime": "10:00 A",
								"endTime": "10:30 A"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2008"
								},
								"example": {
									"success": true,
									"msg": "Time Slot created successfully",
									"data": {
										"deletedAt": null,
										"deletedBy": null,
										"_id": "62ce6bee37e91b0a80fddb2b",
										"startTime": "10:00 A",
										"endTime": "10:30 A",
										"createdAt": "2022-07-13T06:53:34.046Z",
										"updatedAt": "2022-07-13T06:53:34.046Z",
										"__v": 0
									}
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
		"/timeSlot/update/62ce6b4e887df22d203c5ff4": {
			"put": {
				"tags": ["Time Slot"],
				"summary": "Update time Slot",
				"operationId": "UpdatetimeSlot",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatetimeSlotRequest"
							},
							"example": {
								"startTime": "10:10 A",
								"endTime": "10:30 A"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2003"
								},
								"example": {
									"success": true,
									"msg": "Time Slot update successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		},
...
		"/timeSlot/delete/62ce6b4e887df22d203c5ff4": {
			"delete": {
				"tags": ["Time Slot"],
				"summary": "Delete time Slot",
				"operationId": "DeletetimeSlot",
				"parameters": [],
				"responses": {
					"200": {
						"description": "OK",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/m2003"
								},
								"example": {
									"success": true,
									"msg": "Time Slot delete successfully",
									"data": []
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": [
					{
						"bearer": []
					}
				]
			}
		}
	},
	"components": {
		"schemas": {
...
			"Data": {
				"title": "Data",
				"required": [
					"image",
					"status",
					"deletedAt",
					"deletedBy",
					"_id",
					"name",
					"price",
					"time",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"properties": {
					"image": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"deletedAt": {
						"type": "string",
						"nullable": true
					},
					"deletedBy": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "integer",
						"format": "int32"
					},
					"time": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"image": "appointment_book/1657696866655-image (4).png",
					"status": "AVAILABLE",
					"deletedAt": null,
					"deletedBy": null,
					"_id": "62ce726328b7481c283e1c77",
					"name": " Back Pain",
					"price": 700,
					"time": "9:00 AM to 5:00 PM",
					"createdAt": "2022-07-13T07:21:07.024Z",
					"updatedAt": "2022-07-13T07:21:07.024Z",
					"__v": 0
				}
			},
			"m2002": {
				"title": "m2002",
				"required": ["success", "data", "totalCount"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Data"
						},
						"description": ""
					},
					"totalCount": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"success": true,
					"data": [
						{
							"image": "appointment_book/1655721158301-img.png",
							"status": "Available",
							"deletedAt": null,
							"deletedBy": null,
							"_id": "62b046d2d5debb823497dc1c",
							"name": " Back Pain ",
							"price": 700,
							"time": "9:00 AM to 5:00 PM",
							"createdAt": "2022-06-20T10:07:15.090Z",
							"updatedAt": "2022-06-20T10:32:39.125Z",
							"__v": 0
						},
						{
							"image": "appointment_book/1657696907620-image (1).png",
							"status": "Available",
							"deletedAt": null,
							"deletedBy": null,
							"_id": "62ce726328b7481c283e1c77",
							"name": " Back Pain ",
							"price": 700,
							"time": "9:00 AM to 5:00 PM",
							"createdAt": "2022-07-13T07:21:07.024Z",
							"updatedAt": "2022-07-13T07:21:48.467Z",
							"__v": 0
						}
					],
					"totalCount": 2
				}
			},
...
			"m2003": {
				"title": "m2003",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"success": true,
					"msg": "Speciality delete successfully",
					"data": []
				}
			},
...
			"CreatetimeSlotRequest": {
				"title": "CreatetimeSlotRequest",
				"required": ["startTime", "endTime"],
				"type": "object",
				"properties": {
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					}
				},
				"example": {
					"startTime": "10:00 A",
					"endTime": "10:30 A"
				}
			},
			"m2008": {
				"title": "m2008",
				"required": ["success", "msg", "data"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Data4"
					}
				},
				"example": {
					"success": true,
					"msg": "Time Slot created successfully",
					"data": {
						"deletedAt": null,
						"deletedBy": null,
						"_id": "62ce6bee37e91b0a80fddb2b",
						"startTime": "10:00 A",
						"endTime": "10:30 A",
						"createdAt": "2022-07-13T06:53:34.046Z",
						"updatedAt": "2022-07-13T06:53:34.046Z",
						"__v": 0
					}
				}
			},
			"Data4": {
				"title": "Data4",
				"required": [
					"deletedAt",
					"deletedBy",
					"_id",
					"startTime",
					"endTime",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"properties": {
					"deletedAt": {
						"type": "string",
						"nullable": true
					},
					"deletedBy": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"deletedAt": null,
					"deletedBy": null,
					"_id": "62ce6bee37e91b0a80fddb2b",
					"startTime": "10:00 A",
					"endTime": "10:30 A",
					"createdAt": "2022-07-13T06:53:34.046Z",
					"updatedAt": "2022-07-13T06:53:34.046Z",
					"__v": 0
				}
			},
			"UpdatetimeSlotRequest": {
				"title": "UpdatetimeSlotRequest",
				"required": ["startTime", "endTime"],
				"type": "object",
				"properties": {
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					}
				},
				"example": {
					"startTime": "10:10 A",
					"endTime": "10:30 A"
				}
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [],
...
}